"Move along, nothing to see here..."
	Seriously, I made a late-night error back in October of 2000
when I was adding the GPL/LGPL headers to a bunch of files in the
source, and ended up adding the wrong headers to several. I reverted
them back and put the proper LGPL headers in where the libpisock
bindings were used and required (in ./libsock and Java/Perl5
directories). Sorry about that, it's all better now.

	Note: The license for the pisock library, the C++ libary, and
              the Perl, Python, Tcl, and Java bindings have been
              covered under the LGPL since August 23, 1998.