Fixed the fd leak problem with a chainsaw, removing all device duplication code.
Reverted connect.c to the pre-mace version, with some changes.

NOTE!  This changes the pi_accept API, it now returns the sd passed to it, instead of making a duplicate of it and returning that, this will have an impact on all programs which try to treat the 'parent' sd and the 'client' sd differently.

Fixed a libusb reopening bug.

Added the (very hacky) reopen command to dlpsh, to test reopening the link to the PDA.

 9 files changed, 108 insertions(+), 395 deletions(-)