After much discussion, back and forth, modify pilot-xfer's argument handling _again_; it now does:
* Sanity checking before connecting.
* Accepts exactly _one_ of the -bursimfdlL flags and complains
  otherwise.
* -burs still take a <dir>.
* -imfd don't take a particular argument, but apply to all the
  leftovers. This covers cases like

  --install foo.prc --port /dev/pilot
  --install *.prc --port /dev/pilot
  --install --port /dev/pilot --vfsdir /slot1/ *.prc

  You can't mix them, though -- this was deemed too compilcated and
  unused anyway, so no --install *.prc --delete *.mp3.
* -lL don't take any extra arguments, and will complain if they are found.
* --Purge, if present, is done after -bursimfdlL.

I do not claim to have tested all combinations of arguments that
are possible, and there are many that do not even make sense (like
--backup with --vfsdir) that are not complained about.

The --exec option is now left in a strange position. When pilot-xfer
acted upon each argument in turn from left (argv[1]) to right (argv[argc-1]),
doing a system() between two other actions might have made sense.
Now, since we collect arguments and act on them _after_ initial processing
where the position of the options is lost, --exec doesn't have a
specific spot where it should be executed, eg.

        --port /dev/pilot *.prc --exec /bin/halt *.pdb --install

is also a valid command-line, and there's no way to tell that that
exec was "supposed" to happen between the prcs and the pdbs.