And for /dev/pilot, I see DEATH.
Regarding the error message, checking for a null port on a bad result is a
bandaid.  Since I realised pilot_connect would eventually go away, I kinda
painted myself into a corner with my changes to pi_bind.  plu_connect is
basically a copy of pilot_connect with the same flaw, so it gets the same
bandaid.  I need to revisit this problem later, but there's some socket
level stuff that needs to happen first so that we can actually report
useful error messages.