Removed all mention to /dev/pilot in src!  serial:/dev/ttyUSB0 is used instead, just because it's probably a little more obvious that you shouldn't be making random symlinks as newbies keep trying to do for /dev/pilot.
Of course, it means that as soon as the device specs are finished being
hashed out as part of my planned attack on socket code, someone will have
to write some docs to explain it so people don't ask silly questions like
why you would use serial: to talk to a Palm with a USB cradle with Linux
and visor.ko.  I'll worry about that when I'm ready.